Updated: Correct file structure for Text component This is to follow the convention that the component should be under Component/Component.js, but using imports with just Component implies adding a Component/index.js
