+++ src/Badge/Badge.js
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import { View } from 'react-native';
 import { defaultTokens } from '@kiwicom/orbit-design-tokens';
-import Text from '../Text';
+import { Text } from '../Text';
 import StyleSheet from '../PlatformStyleSheet';
 import type { BadgeProps, BadgeType } from './BadgeTypes';
 import { textColor, wrapperColor } from './styles';
