+++ src/Button/ButtonTitle.js
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import StyleSheet from '../PlatformStyleSheet';
 
-import Text from '../Text';
+import { Text } from '../Text';
 import { textColor } from './styles';
 
 import type { StylePropType } from '../PlatformStyleSheet/StyleTypes';
