+++ src/Icon/IconsList.js
@@ -4,7 +4,7 @@ import * as React from 'react';
 import { FlatList, View } from 'react-native';
 
 import { Icon } from '.';
-import Text from '../Text';
+import { Text } from '../Text';
 import StyleSheet from '../PlatformStyleSheet';
 import iconsMap from './icons.json';
