+++ src/LocalizedPrice/LocalizedPrice.js
@@ -2,7 +2,7 @@
 
 import * as React from 'react';
 import { defaultTokens } from '@kiwicom/orbit-design-tokens';
-import Text from '../Text';
+import { Text } from '../Text';
 import StyleSheet from '../PlatformStyleSheet';
 import type { LocalizedPriceProps } from './LocalizedPriceTypes';
