+++ src/Notification/InformativeNotification.js
@@ -5,7 +5,7 @@ import { defaultTokens } from '@kiwicom/orbit-design-tokens';
 import { Animated } from 'react-native';
 
 import StyleSheet from '../PlatformStyleSheet';
-import Text from '../Text';
+import { Text } from '../Text';
 
 import type { StylePropType } from '../PlatformStyleSheet/StyleTypes';
