+++ src/Rating/Rating.js
@@ -1,7 +1,7 @@
 // @flow
 
 import * as React from 'react';
-import Text from '../Text';
+import { Text } from '../Text';
 import StyleSheet from '../PlatformStyleSheet';
 import type { StylePropType } from '../PlatformStyleSheet/StyleTypes';
